readjust the traffic control module
