+++ constants/common.go
@@ -1,0 +1,13 @@
+package constants
+
+// Traffic Control
+const (
+	TrafficControlTaskControlTypePercent       = "Percent"
+	TrafficControlTaskControlTypeQuantity      = "Quantity"
+	TrafficControlTaskControlGranularityGlobal = "Global"
+	TrafficControlTaskControlGranularitySingle = "Single"
+	TrafficControlTaskControlLogicGuaranteed   = "Guaranteed"
+	TrafficControlTaskControlLogicApproach     = "Approach"
+	TrafficControlTargetStatisPeriodDaily      = "Daily"
+	TrafficControlTargetStatisPeriodHour       = "Hour"
+)
