+++ recconf/recconf.go
@@ -247,8 +247,10 @@ type AlgoConfig struct {
 }
 
 type PIDControllerConfig struct {
-	SyncPIDStatus          bool
 	HologresName           string
+	HologresTableName      string
+	PrimaryKey             string
+	SyncPIDStatus          bool
 	LoadItemFeature        bool
 	PreloadItemFeature     bool
 	AllocateExperimentWise bool
@@ -256,7 +258,7 @@ type PIDControllerConfig struct {
 	MaxItemCacheTime       int
 	RedisName              string
 	RedisKeyPrefix         string
-	SampleTime             float32
+	TimeWindow             int
 	DefaultKp              float32
 	DefaultKi              float32
 	DefaultKd              float32
